Apply changes for progess monitor
